\name{test.net}
\alias{test.net}
\docType{data}
\title{Netmodels sample data}
\description{
  This graph object is designed to provide an standard object to test the different functions of this package.
}
\usage{data(test.net)}
\format{
  The format is:
List of 9
 $ : num 89
 $ : logi FALSE
 $ : num [1:1332] 41 41 42 43 44 44 44 45 46 46 ...
 $ : num [1:1332] 0 1 0 2 1 1 3 3 2 2 ...
 $ : num [1:1332] 1282  839  396 1189  746 ...
 $ : num [1:1332] 1282  839  396 1273 1272 ...
 $ : num [1:90] 0 0 0 6 21 27 81 162 168 180 ...
 $ : num [1:90]   0 364 491 618 702 ...
 $ :List of 4
  ..$ : num [1:2] 1 0
  ..$ : list()
  ..$ :List of 1
  .. ..$ id: chr [1:89] " node 1 " " node 2 " " node 3 " " node 4 " ...
  ..$ : list()
 - attr(*, "class")= chr "igraph"
}
\details{
Igraph graph object
}
\source{
}
\references{
See the Igraph library documentation for the object details
}
\examples{
data(test.net)
}
\keyword{datasets}
